[IMPROVE] Admin ui step 1 (#13393)
Just few changes to update admin pages
