+++ packages/rocketchat-integrations/client/stylesheets/integrations.css
@@ -53,10 +53,6 @@
 	}
 }
 
-.zapier-btn {
-	margin-right: 5px;
-}
-
 .integrate-other-ways {
 	& p {
 		font-size: 1rem;
