+++ packages/rocketchat-ui-admin/client/users/adminUsers.js
@@ -11,6 +11,10 @@ import s from 'underscore.string';
 import { RocketChatTabBar } from 'meteor/rocketchat:lib';
 
 Template.adminUsers.helpers({
+	searchText() {
+		const instance = Template.instance();
+		return instance.filter && instance.filter.get();
+	},
 	isReady() {
 		const instance = Template.instance();
 		return instance.ready && instance.ready.get();
@@ -40,6 +44,26 @@ Template.adminUsers.helpers({
 			data: Template.instance().tabBarData.get(),
 		};
 	},
+	onTableScroll() {
+		const instance = Template.instance();
+		return function(currentTarget) {
+			if (
+				currentTarget.offsetHeight + currentTarget.scrollTop >=
+				currentTarget.scrollHeight - 100
+			) {
+				return instance.limit.set(instance.limit.get() + 50);
+			}
+		};
+	},
+	// onTableItemClick() {
+	// 	const instance = Template.instance();
+	// 	return function(item) {
+	// 		Session.set('adminRoomsSelected', {
+	// 			rid: item._id,
+	// 		});
+	// 		instance.tabBar.open('admin-room');
+	// 	};
+	// },
 });
 
 Template.adminUsers.onCreated(function() {
