+++ packages/rocketchat-ui-admin/server/publications/adminRooms.js
@@ -21,6 +21,7 @@ Meteor.publish('adminRooms', function(filter, types, limit) {
 			cl: 1,
 			u: 1,
 			usernames: 1,
+			usersCount: 1,
 			muted: 1,
 			ro: 1,
 			default: 1,
