+++ packages/rocketchat-ui/client/index.js
@@ -86,6 +86,8 @@ import './components/selectDropdown.html';
 import './components/selectDropdown';
 import './components/header/header.html';
 import './components/header/header';
+import './components/header/headerRoom.html';
+import './components/header/headerRoom';
 import './components/contextualBar.html';
 import './components/contextualBar';
 export { getAvatarUrFromUsername } from 'meteor/rocketchat:utils';
