+++ packages/rocketchat_theme/client/imports/components/header.css
@@ -1,24 +1,15 @@
-.rc-header.rc-header--fullpage {
-	padding: 1.25rem 1rem 0.75rem;
-
-	& .rc-header__wrap {
-
-		padding: 0.75rem 0;
-
-		box-shadow: none;
-	}
-
-	& .rc-header__block {
-		font-size: 1.375rem;
-	}
-
-	& .rc-button {
-		margin: -10px 0;
-	}
-}
-
 .rc-header {
-	padding: 0 0.5rem;
+	padding: 0.5rem;
+
+	font-size: var(--text-heading-size);
+
+	&--room {
+		padding: 1.25rem;
+
+		box-shadow: 0 1px 2px 0 rgba(31, 35, 41, 0.08);
+
+		font-size: var(--header-title-font-size);
+	}
 
 	&__wrap {
 		z-index: 2;
@@ -27,30 +18,26 @@
 
 		flex: 0 0 auto;
 
-		margin: 0 -0.5rem;
+		height: 36px;
 
-		padding: var(--header-padding);
+		margin: 0 -0.5rem;
 
 		white-space: nowrap;
 
 		background-color: var(--header-background-color);
-		box-shadow: 0 1px 2px 0 rgba(31, 35, 41, 0.08);
-
-		font-size: var(--header-title-font-size);
 
 		align-items: center;
 
 		justify-content: space-between;
 	}
 
-	&--height-fixed {
-		height: 78px;
-	}
-
 	&__block {
 		display: flex;
 
-		margin: 0 0.5rem;
+		margin: 0 -0.5rem;
+
+		padding: 0 0.5rem;
+
 		align-items: center;
 	}
 
@@ -185,9 +172,10 @@
 	}
 
 	&__toggle-favorite {
-		color: var(--header-toggle-favorite-star-color);
 
-		font-size: 1.5rem;
+		padding: 0 0.25rem;
+
+		color: var(--header-toggle-favorite-star-color);
 
 		&.empty {
 			color: var(--header-toggle-favorite-color);
@@ -198,7 +186,7 @@
 		}
 
 		& > .rc-header__icon {
-			font-size: 1.5rem;
+			font-size: 2rem;
 		}
 
 		&:hover {
@@ -213,8 +201,14 @@
 	&__image {
 		width: 32px;
 		height: 32px;
-		margin: 0 0.5rem;
 		flex-shrink: 0;
+	}
+
+	.rc-button {
+		height: 36px;
+		min-height: 36px;
+
+		margin: 0 0.25rem;
 	}
 }
