+++ packages/rocketchat_theme/client/imports/components/tabs.css
@@ -9,7 +9,7 @@
 
 	width: 100%;
 
-	margin: 0 -1rem -2px -1rem;
+	margin: 0 -1rem -2px;
 }
 
 .tab {
