+++ packages/rocketchat_theme/client/imports/general/base_old.css
@@ -192,17 +192,24 @@
 	top: 0;
 	left: 0;
 
+	display: flex;
+
 	overflow-y: hidden;
+	flex-direction: column;
 
 	width: 100%;
 	height: 100%;
 
-	& .content {
-		overflow-y: scroll;
+	padding: 1.25rem 2rem;
 
-		width: 100%;
-		height: calc(100% - 60px);
-		padding: 25px 40px;
+	& .content {
+		display: flex;
+		overflow-y: auto;
+		flex-direction: column;
+		flex: 1 1 auto;
+
+		margin: 0 -2rem;
+		padding: 2rem 2rem 0;
 
 		line-height: 1.3em;
 		-webkit-overflow-scrolling: touch;
@@ -264,28 +271,6 @@
 					text-align: center;
 				}
 			}
-		}
-	}
-
-	& table {
-		overflow: hidden;
-
-		width: 100%;
-		margin-bottom: 30px;
-
-		& th,
-		& td {
-			padding: 0.6rem 0.7rem;
-
-			user-select: text;
-			text-align: left;
-			vertical-align: middle;
-
-			border-width: 0 0 1px;
-		}
-
-		& th {
-			white-space: nowrap;
 		}
 	}
 }
@@ -1790,10 +1775,16 @@ rc-old select,
 	& .section {
 		padding: 20px 0;
 
-		border-bottom: 2px solid #dddddd;
+		&:not(:only-child) {
+			border-bottom: 2px solid #dddddd;
 
-		&.section-collapsed .section-content {
-			display: none;
+			&.section-collapsed .section-content {
+				display: none;
+			}
+
+			.section-title-right {
+				visibility: visible;
+			}
 		}
 	}
 
@@ -1813,8 +1804,12 @@ rc-old select,
 			flex-grow: 1;
 		}
 
-		& .section-title-right > .rc-button {
-			font-size: 1.25rem;
+		& .section-title-right {
+			visibility: hidden;
+
+			& > .rc-button {
+				font-size: 1.25rem;
+			}
 		}
 	}
 
@@ -2054,8 +2049,6 @@ rc-old select,
 	& .results {
 		margin-bottom: 10px;
 		padding: 10px 0;
-
-		border-width: 0 0 1px;
 
 		font-weight: 300;
 
@@ -3505,9 +3498,19 @@ body:not(.is-cordova) {
 	z-index: 2;
 
 	display: flex;
-	flex: 0 0 41px;
+	flex: 0 0 40px;
 
-	border-width: 0 0 0 1px;
+	transform: box-shadow 0.3s;
+
+	box-shadow: -1px 0 0 1px #cccccc26;
+
+	&.opened {
+		box-shadow: -1px 0 5px 2px #cccccc26;
+
+		& > .flex-tab-bar {
+			box-shadow: -1px 0 5px 2px #cccccc26;
+		}
+	}
 
 	& .flex-tab {
 		position: relative;
@@ -3736,7 +3739,9 @@ body:not(.is-cordova) {
 
 		min-width: 40px;
 
-		border: 1px solid rgba(31, 35, 41, 0.08);
+		transition: box-shadow 0.3s;
+
+		box-shadow: -1px 0 0 1px #cccccc26;
 
 		& .tab-button {
 			position: relative;
@@ -4004,7 +4009,7 @@ body:not(.is-cordova) {
 	}
 
 	& .contact-code {
-		margin: -5px 0 10px 0;
+		margin: -5px 0 10px;
 
 		font-size: 12px;
 	}
@@ -5605,8 +5610,6 @@ body:not(.is-cordova) {
 		right: 40px;
 
 		height: 100%;
-
-		border-width: 0 0 0 1px;
 	}
 }
