+++ packages/rocketchat_theme/server/colors.less
@@ -247,12 +247,6 @@
  * Document components
  */
 
-.page-container {
-	tr:hover td {
-		background-color: @content-background-color;
-	}
-}
-
 .burger i {
 	background-color: @primary-font-color;
 }
